Refactor move TextField to dialog
